Move conformance tests to tests directory
