+++ .github/workflows/release-pr.yml
@@ -74,8 +74,6 @@ jobs:
           sed -i -e "s/v${{ steps.vars.outputs.current_version }}/v${{ inputs.version }}/g" README.md
           sed -i -e "s/\[${{ steps.vars.outputs.current_version }}\]/\[${{ inputs.version }}\]/g" README.md
           sed -i -e "s/VERSION = edge/VERSION = ${{ inputs.version }}/g" Makefile
-          sed -i -e "s/VERSION =  ${{ steps.vars.outputs.current_version }}/VERSION = ${{ inputs.version }}/g" conformance/Makefile
-          sed -i -e "s/TAG = edge/TAG = ${{ inputs.version }}/g" **/Makefile
           sed -i "6r .github/CHANGELOG_TEMPLATE.md" CHANGELOG.md
           sed -i -e "s/%%VERSION%%/${{ inputs.version }}/g" CHANGELOG.md
           make generate-manifests
