+++ .gitignore
@@ -12,7 +12,7 @@
 *.out
 cover.html
 cmd-cover.html
-conformance/conformance-profile.yaml
+conformance-profile.yaml
 
 # Dependency directories (remove the comment below to include it)
 # vendor/
