+++ config/tests/static-deployment.yaml
