+++ docs/developer/testing.md
@@ -103,4 +103,4 @@ submitting them for review and integration into the project.
 
 ## Gateway API Conformance Testing
 
-To run Gateway API conformance tests, please follow the instructions on [this](/conformance/README.md) page.
+To run Gateway API conformance tests, please follow the instructions on [this](/tests/README.md#conformance-testing) page.
