+++ tests/Dockerfile
@@ -1,0 +1,3 @@
+# syntax=docker/dockerfile:1.7
+# this is here so we can grab the latest version of kind and have dependabot keep it up to date
+FROM kindest/node:v1.29.2
