+++ tests/conformance/Dockerfile
@@ -1,0 +1,10 @@
+# syntax=docker/dockerfile:1.7
+FROM golang:1.22
+
+WORKDIR /go/src/github.com/nginxinc/nginx-gateway-fabric/tests/conformance
+
+COPY --link go.mod /go/src/github.com/nginxinc/nginx-gateway-fabric/tests/
+COPY --link go.sum /go/src/github.com/nginxinc/nginx-gateway-fabric/tests/
+RUN go mod download
+
+COPY --link conformance /go/src/github.com/nginxinc/nginx-gateway-fabric/tests/conformance
