+++ tests/scripts/install-gateway.sh
