+++ tests/scripts/uninstall-gateway.sh
